v1.01 Read to port to birdview-3, needs fine tuning
Everything works, including the needlessly complex individual cage monitoring system
I would like to add some diagnostics and fail conditions so that if something weird happens
we still get good data. Here are some cases to watch out for:
-Camera fails to get tag so it has to revert to whole motion tracking
-One of the cameras is unplugged
-Birds just never stop moving

Most of that framework is in place, so I just need to add it in.
But by all accounts this is looking very good, and needs some fine tuning on the system that will actually run it.